refactor: Add Mark as Watched Button to Watch Page
Make the button now a component, then add it to Watch Page.
