+++ front-end/app/media/[id]/components/GoGoAnimeEpisodeContainer/component.module.css
@@ -68,43 +68,3 @@
     text-decoration: underline;
 
 }
-
-.container .title_button_container button {
-
-    display: flex;
-    background: transparent;
-    border: 0;
-
-    border-radius: 50%;
-
-    padding: 4px;
-
-}
-
-.container .title_button_container button[data-active=false]:hover {
-    background: var(--white-25);
-}
-
-.container .title_button_container button[data-active=false]:hover svg {
-    fill: var(--brand-color);
-}
-
-.container .title_button_container button[data-active=true]:hover {
-    background: var(--black-100);
-}
-
-.container .title_button_container button[data-active=true]:hover svg {
-    fill: var(--white-75);
-}
-
-.container .title_button_container button[data-active=true] svg {
-
-    fill: var(--brand-color);
-
-}
-
-.container .title_button_container button[data-active=false] svg {
-
-    fill: var(--white-75);
-
-}
